% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractor.R
\name{extractor}
\alias{extractor}
\title{Extract information on the model fits from a JSON package as data frames}
\usage{
extractor(fin)
}
\arguments{
\item{fin}{name of the input JSON file}
}
\value{
a list of data frames.  Each frame is ordered by deviance (rounded to 2 decimal places)
   and by elapsed time withing deviance.
}
\description{
Extract information on the model fits from a JSON package as data frames
}

