% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table_maker.R
\name{jsondir}
\alias{jsondir}
\title{From a directory name extract all the timings from the JSON files}
\usage{
jsondir(dname)
}
\arguments{
\item{dname}{name of a firectory}
}
\value{
a list of data sets, models within dataset, and fits within models
}
\description{
Given a directory name read all the files in the directory
  with names ending in .json.  Returns a list of lists.  The names of
  the list.
}
\examples{
str(jsondir(system.file("JSON",package="Timings")),4)
}

