% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retime.R
\name{retime}
\alias{retime}
\title{Retime a model fit}
\usage{
retime(lst, ...)
}
\arguments{
\item{lst}{a list containing, at least, components \code{formula} and \code{data}}
}
\value{
the same list augmented with a new timing
}
\description{
\code{retime} adds a new timing to a list, usually created from a JSON file,
by refitting the model.
}

